requirements.txt
fastapi
uvicorn
pydantic
httpx
openai
python-dotenv
